requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
autobahn==25.9.1
Automat==25.4.16
certifi==2025.4.26
cffi==2.0.0
charset-normalizer==3.4.1
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
dataclasses-json==0.6.7
distro==1.9.0
dj-database-url==2.3.0
Django==5.2
django-eventstream==5.3.2
django-grip==3.5.2
django-picklefield==3.3
django-q2==1.8.0
frozenlist==1.6.0
greenlet==3.2.1
gripcontrol==4.2.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-community==0.3.23
langchain-core==0.3.58
langchain-openai==0.3.16
langchain-text-splitters==0.3.8
langsmith==0.3.42
MarkupSafe==3.0.3
marshmallow==3.26.1
multidict==6.4.3
mypy_extensions==1.1.0
numpy==2.2.5
openai==1.76.0
orjson==3.10.18
packaging==24.2
pillow==11.2.1
pluggy==1.6.0
propcache==0.3.1
psycopg2-binary==2.9.10
pubcontrol==3.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
retell-sdk==4.28.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlparse==0.5.3
square==0.0.3
squareup==42.0.0.20250416
stripe==12.1.0
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
twilio==9.5.2
Twisted==25.5.0
txaio==25.9.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
whitenoise==6.9.0
yarl==1.20.0
zope.interface==8.0.1
zstandard==0.23.0
